Add files via upload
The dataset we have been working for is (Absenteeism from workplace). 
In the project we cleaned and preprocessed data for further analysis. On the project, among many tasks, one of the important challenge was dividing the 29 unique reasons of absence to 4 different categories, which are (reason_1, reason_2, reason_3 and reason_4). More than that we dig more deeper for each column so there wouldn`t be any misleading for analysis